% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.number.at.risk.for.HD}
\alias{compute.number.at.risk.for.HD}
\title{compute.number.at.risk.for.HD: This function returns number of subjects who are at risk for HD for each study at times.}
\usage{
compute.number.at.risk.for.HD(
  study.names,
  data.file.names,
  event.outcome.names,
  nonfunctional.covariate.names,
  functional.covariate.names,
  nonfunctional.covariate.value,
  functional.covariate.values.of.interest,
  time.points.for.prediction,
  estimated.parameters.common.for.all.studies
)
}
\arguments{
\item{study.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{data.file.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{event.outcome.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.value}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.prediction}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{estimated.parameters.common.for.all.studies}{See this argument in the \code{\link{jprat.wrapper}} function.}
}
\value{
a table for the number of subjects whoare at risk for HD for each study at times.(ignores specific covariate values)
}
\description{
compute.number.at.risk.for.HD: This function returns number of subjects who are at risk for HD for each study at times.
}
