% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert.cag}
\alias{convert.cag}
\title{This function converts the uniformly distributed values (CAG repeats length) on [0,1] to the original values (CAG repeats length from real data).}
\usage{
convert.cag(x, xmin = xmin, xmax = xmax)
}
\description{
This function converts the uniformly distributed values (CAG repeats length) on [0,1] to the original values (CAG repeats length from real data).
}
