% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert.new.notation.to.old.for.get.results}
\alias{convert.new.notation.to.old.for.get.results}
\title{convert.new.notation.to.old.for.get.results: This code converts the new notation into the original code notation for getting results (plots)}
\usage{
convert.new.notation.to.old.for.get.results(
  study.names,
  time.points.for.prediction,
  time.points.for.conditional.prediction,
  time.points.for.conditional.prediction.toadd,
  nonfunctional.covariate.names,
  nonfunctional.covariate.values.for.prediction,
  functional.covariate.names,
  functional.covariate.values.of.interest,
  functional.covariate.values.of.interest.ci,
  functional.covariate.values.for.prediction,
  xmin,
  xmax,
  event.outcome.names,
  use_real_data,
  use.functional.beta.intercept,
  use.functional.event.coefficients,
  use.functional.study.coefficients,
  number.of.bootstraps,
  what.analyzed.separately,
  estimated.parameters.common.for.all.studies,
  glm.link.type,
  use.bootstrap.variance,
  time.points.of.interest,
  time.points.of.interest.ci,
  label.for.alpha.values.over.time,
  nonfunctional.covariate.comparisons,
  add.number.at.risk.legend,
  ylabel.for.plots.comparing.studies,
  xlabel.for.plots.comparing.studies,
  plot.confidence.intervals,
  color.labels,
  legend.labels,
  event.comparison.table,
  functional.covariate.comparisons,
  functional.covariate.comparisons.for.sample.size
)
}
\arguments{
\item{study.names}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.prediction}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.conditional.prediction}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.conditional.prediction.toadd}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.values.for.prediction}{See the argument in the \code{\link{convert.new.notation.to.old.for.jprat}} function.}

\item{functional.covariate.names}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest.ci}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.for.prediction}{See the argument in the \code{\link{convert.new.notation.to.old.for.jprat}} function.}

\item{xmin}{See the argument in the \code{\link{convert.new.notation.to.old.for.jprat}} function.}

\item{xmax}{See the argument in the \code{\link{convert.new.notation.to.old.for.jprat}} function.}

\item{event.outcome.names}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{use_real_data}{See the argument in the \code{\link{convert.new.notation.to.old.for.jprat}} function.}

\item{use.functional.beta.intercept}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{use.functional.event.coefficients}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{use.functional.study.coefficients}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{number.of.bootstraps}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{what.analyzed.separately}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{estimated.parameters.common.for.all.studies}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{glm.link.type}{a character value for the name of the link function for the additive mixed effect model;
: "logit” for proportional odds model or "cloglog" for cox proportional hazards.
Default is "logit".}

\item{use.bootstrap.variance}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.of.interest}{ADD DETAILS HERE!!}

\item{time.points.of.interest.ci}{a vector of time points at which confidence intervals of
\eqn{\alpha(X,t)} are predicted. These time points are used to be labeled for the plot.}

\item{label.for.alpha.values.over.time}{a vector of specific time points at which the smooth functional parameters \eqn{\alpha(X,t)} will be predicted. These time points are used to be labeled for the plot.}

\item{nonfunctional.covariate.comparisons}{a list of comparison sets where the predictions for nonfunctional covariates Z will be compared. Default is no comparison made for the type of analysis "7c" and "7c-converters".}

\item{add.number.at.risk.legend}{See the argument in the \code{\link{view.all.results}} function.}

\item{ylabel.for.plots.comparing.studies}{See the argument in the \code{\link{view.all.results}} function.}

\item{xlabel.for.plots.comparing.studies}{See the argument in the \code{\link{view.all.results}} function.}

\item{plot.confidence.intervals}{ADD DETAILS HERE!}

\item{color.labels}{See the argument in the \code{\link{view.all.results}} function.}

\item{legend.labels}{See the argument in the \code{\link{view.all.results}} function.}

\item{event.comparison.table}{a list of values corresponding to the event of interest outcomes. Default is a list of nc1, which contains a vector of values 1, 2, 3, 4.}

\item{functional.covariate.comparisons}{See the argument in the \code{\link{view.all.results}} function.}

\item{functional.covariate.comparisons.for.sample.size}{See the argument in the \code{\link{view.all.results}} function.}
}
\value{
FILL IN HERE!!
}
\description{
convert.new.notation.to.old.for.get.results: This code converts the new notation into the original code notation for getting results (plots)
}
