% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert.new.notation.to.old.for.jprat}
\alias{convert.new.notation.to.old.for.jprat}
\title{convert.new.notation.to.old.for.jprat: This code converts the new notation into the original code notation for JPRAT analysis.}
\usage{
convert.new.notation.to.old.for.jprat(
  study.names,
  data.sets.as.list,
  time.points.for.prediction,
  time.points.for.conditional.prediction,
  time.points.for.conditional.prediction.toadd,
  nonfunctional.covariate.names,
  nonfunctional.covariate.values.for.prediction,
  functional.covariate.names,
  functional.covariate.values.for.prediction,
  xmin,
  xmax,
  othercovariate.names,
  event.outcome.names,
  delta.names,
  use_real_data,
  use.functional.beta.intercept,
  use.functional.event.coefficients,
  use.functional.study.coefficients,
  number.of.bootstraps,
  check.study.equality,
  what.analyzed.separately,
  estimated.parameters.common.for.all.studies,
  estimation.when.censoring.depends.on.z,
  glm.link.type,
  use.bootstrap.variance,
  clusters.are.families
)
}
\description{
convert.new.notation.to.old.for.jprat: This code converts the new notation into the original code notation for JPRAT analysis.
}
