% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{criteria.time.points.for.jprat.analysis}
\alias{criteria.time.points.for.jprat.analysis}
\title{criteria.time.points.for.jprat.analysis:
This function returns warning message if the criteria are not met for \code{functional.covariate.values.of.interest}, \code{functional.covariate.values.of.interest.ci},
\code{time.points.of.interest}, \code{time.points.of.interest.ci}, \code{functional.covariate.comparisons} and \code{time.points.for.conditional.prediction}.}
\usage{
criteria.time.points.for.jprat.analysis(
  study.names,
  input.data.list,
  time.points.for.prediction,
  nonfunctional.covariate.names,
  functional.covariate.names,
  nonfunctional.covariate.value,
  time.points.for.conditional.prediction,
  functional.covariate.values.of.interest,
  functional.covariate.values.of.interest.ci,
  functional.covariate.comparisons,
  time.points.of.interest,
  time.points.of.interest.ci
)
}
\arguments{
\item{study.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{input.data.list}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.prediction}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.value}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.conditional.prediction}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest.ci}{See this argument in the \code{\link{view.all.results}} function.}

\item{functional.covariate.comparisons}{See this argument in the \code{\link{view.all.results}} function.}

\item{time.points.of.interest}{See this argument in the \code{\link{view.all.results}} function.}

\item{time.points.of.interest.ci}{See this argument in the \code{\link{view.all.results}} function.}
}
\value{
warning messages if the criteria are not met for arguments  \code{functional.covariate.values.of.interest}, \code{functional.covariate.values.of.interest.ci},
\code{time.points.of.interest}, \code{time.points.of.interest.ci}, \code{functional.covariate.comparisons} and \code{time.points.for.conditional.prediction}.
}
\description{
criteria.time.points.for.jprat.analysis:
This function returns warning message if the criteria are not met for \code{functional.covariate.values.of.interest}, \code{functional.covariate.values.of.interest.ci},
\code{time.points.of.interest}, \code{time.points.of.interest.ci}, \code{functional.covariate.comparisons} and \code{time.points.for.conditional.prediction}.
}
