% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data.reformatted.for.analysis.results}
\alias{data.reformatted.for.analysis.results}
\title{Reformat Datasets}
\usage{
data.reformatted.for.analysis.results(
  study.names,
  event.outcome.names = NULL,
  color.names,
  legend.names,
  which.nonfunctional.covariate.comparisons = NULL
)
}
\arguments{
\item{study.names}{A character vector of name of studies used in the analysis. e.g., c("cohort", "predict", "pharos").}

\item{event.outcome.names}{A character vector for the name of the time-to-event of interest (outcomes).}

\item{color.names}{A character vector to denote which colors will be used for the labels of events of interest in plots. e.g., c("firebrick1", "darkgreen", "black").}

\item{legend.names}{A character vector to label event of interests in plots. e.g., c("Motor Diagnosis (DCL=4)",  "Cognitive Impairment" , "Stage II TFC").}

\item{which.nonfunctional.covariate.comparisons}{A numeric vector for which nonfunctional covariates will be compared.
In our analysis, we only consider one nonfunctional covariate for baseline age. The default value is c(1, 1).
We only consider one nonfunctional covariate and there is nothing to compare between covariates.}
}
\value{
A list of
     \item{num_study}{The number of studies used in analyses. If the real data analysis used three studies called "cohort", "predict", "pharos", then number of studies is 3. i.e., \code{num_study=3}.}
     \item{nonfunctional.covariate.comparisons}{A list of comparison sets where the predicted values of nonfunctional covariates \eqn{Z} will be compared.}
     \item{color.labels}{A character vector of the colors' name used in labeling events of interest in plots. Users need to provide the color names.}
     \item{legend.labels}{A character vector to label event names in plots. Users need to provide the names for events of interest.}
    \item{event.comparison.table}{A list of numbers corresponding to outcomes.}
}
\description{
This function reforms datasets so that JPRAT can understand and proceed analysis.
}
