% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_setting_estimation_results.R
\name{default.options.for.data.setting}
\alias{default.options.for.data.setting}
\title{default.options.for.data.setting}
\usage{
default.options.for.data.setting()
}
\value{
this function returns default parameters setting to run JPRAT on one data set.
}
\description{
default.options.for.data.setting
}
