% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDData_Cohort_Predict_Pharos.R
\docType{data}
\name{ex_data_cohort}
\alias{ex_data_cohort}
\title{Generated data set for the Huntington's disease (HD) study: Cohort(n=430)}
\format{
A dataframe including 430 subjects' longitudinal data with the following 12 variables:
\describe{
  \item{CAG}{CAG repeats length.}
  \item{base_age}{ages at baseline, when subjects visited clinics for the first time.}
  \item{hdage_nobase}{ages when HD diagnosis has occurred for subjects who did not have HD diagnosis at the baseline.}
  \item{mcione}{ages when the first multi-domain cognitive impairments has occurred for subjects.}
  \item{dep2}{ages when depression were occurred for subjects.}
  \item{delta.hdage_nobase}{censoring indicator for individuals who entered the study without a diagnosis (1 if a person had a HD diagnosis, 0 otherwise).}
  \item{delta.mcione}{censoring indicator whether individuals have experienced the first cognitive impairments (1 if a person has experienced cognitive impairment at first, 0 otherwise).}
  \item{delta.dep2}{censoring indicator whether individuals have experienced depressions (1 if a person has experienced depression, 0 otherwise).}
  \item{delta.tfctwo}{censoring indicator whether individuals have experienced reduction of occupational capacity (1 if a person has experienced on reduction of occupational capacity, 0 otherwise).}
  \item{first year}{year when patients visits clinics for the first time.}
  \item{last year}{year when patients stop visiting clinics.}
}
}
\source{
see data-raw/data_generation.R
}
\usage{
ex_data_cohort
}
\description{
Generated data set for the Huntington's disease (HD) study: Cohort(n=430)
}
\keyword{datasets}
