% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gamm4.estimates}
\alias{gamm4.estimates}
\title{gamm4.estimates: used to get ``gamm4" estimate}
\usage{
gamm4.estimates(
  arbitrary,
  num_study,
  np,
  count.store,
  data,
  num_time,
  time_val,
  time_choice.predicted,
  time_choice.predicted.toadd,
  count.store.outside,
  n,
  nmax,
  m,
  maxm,
  la,
  lb,
  xks,
  truth,
  num_xx,
  knot.length,
  real_data,
  family.data,
  zeval,
  z.choice,
  param.label,
  beta0int,
  gamma.param,
  omega.param,
  spline.constrain,
  common.param.estimation,
  est.cens.depend.z,
  par_fu,
  analyze.separately,
  link.type,
  null.theta,
  z.proportions
)
}
\arguments{
\item{arbitrary}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{num_study}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{np}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{count.store}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{data}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{num_time}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{time_val}{See the argument in the \code{\link{all_null_theta}} function.}

\item{time_choice.predicted}{See the argument in the \code{\link{all_null_theta}} function.}

\item{time_choice.predicted.toadd}{See the argument in the \code{\link{all_null_theta}} function.}

\item{count.store.outside}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{n}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{nmax}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{m}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{maxm}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{la}{See the argument in the \code{\link{all_null_theta}} function.}

\item{lb}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{xks}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{truth}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{num_xx}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{knot.length}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{real_data}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{family.data}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{zeval}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{z.choice}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{param.label}{See the argument in the \code{\link{all_null_theta}} function.}

\item{beta0int}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{gamma.param}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{omega.param}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{spline.constrain}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{common.param.estimation}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{est.cens.depend.z}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{par_fu}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{analyze.separately}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{link.type}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{null.theta}{See the argument in the \code{\link{all_null_theta}} function.}

\item{z.proportions}{See the argument in the \code{\link{jprat.main.estimates}} function.}
}
\value{
A list of

\item{betaest}{the array of estimates for the functional coefficient \eqn{\beta(t)} (beta0, beta1)
               with its estimates (est), estimated variances (varest), estimated lower bounds (varlo) and upper bounds (varhi)  of confidence intervals at each time point for the event of interest per study, where the dimensions of the array are the number of studies (num\_study), the number of clinical events per study (np), the length of the vector of time points(num\_time), the length of parameter label (param.label) and the length of the name for the values (val="est", "varest", "varlo", "varhi").
               See the argument \code{null.theta.est.ci} in \code{\link{all_null_theta}} for the structure of array.}
\item{alphasest}{the array of estimates for the smooth functional coefficient \eqn{\alpha(X,t)} (alphas)
                 with its estimates (est), estimated variances (varest), estimated lower bounds (varlo)
                 and upper bounds (varhi) of confidence intervals at specific value of x and each time point
                 for the event of interest per study, where the dimensions of the array are the number of studies (\code{num_study}),
                 the number of clinical events per study (np), the length of the functional covariate values of X (\code{num_xx}),
                 the length of the vector of time points(\code{num_time}) and the length of smooth functional parameters \eqn{\alpha(X,t)} (la)
                 and the length of the name for the values (val="est", "varest", "varlo", "varhi").
                 See the argument \code{null.theta.est.ci} in the \code{\link{all_null_theta}} function for the structure of array.}
\item{count.store}{See this argument in the \code{\link{jprat.wrapper}} function.}
\item{count.store.outside}{See this argument in the \code{\link{jprat.wrapper}} function.}
\item{Ftest}{array of estimates for the monotone marginal distribution \eqn{F_{es}(t|X, Z)} (Ft) with its estimates (est),
            estimated variances (varest), estimated lower bounds (varlo) and upper bounds (varhi) of confidence intervals at specific value of covariates Z and X,
            and at each time point for the event of interest per study, where the dimensions of the array are the number of studies (\code{num_study}),
            the number of clinical events per study (\code{np}), the length of nonfunctional covariate values Z (\code{z.choice}),
            the length of functional covariate values X (\code{num_xx}) and
            the length of the name for the values (val="est", "varest", "varlo", "varhi").
            See the argument \code{null.theta.est.ci} in the \code{\link{all_null_theta}} function for the structure of array.}
\item{Ftest.predicted}{array of the predicted values for the monotonic marginal distribution \eqn{F_{es}(t|X, Z, t>t_0)} (Ft)
                      beyond time \eqn{t_0} at each time point for each event of interest per study,
                      where the dimensions of the array are the number of studies (\code{num_study}),
                      the number of clinical events per study (\code{np}),
                      the length of nonfunctional covariate values Z (\code{z.choice}),
                      the length of functional covariate values X (\code{num_xx}),
                      the predicted time points (\code{time_choice.predicted.toadd} is added to
                      \code{time_choice.predicted}), the choice of time points for prediction
                      (\code{time_choice.predicted}), and the length of the name for the values (val="est", "varest", "varlo", "varhi").
                     See the argument \code{null.theta.est.ci} in the \code{\link{all_null_theta}} function for the structure of array.
                     If there exists some time points for the prediction (\code{time_choice.predicted}), this value exists.
                     See the argument \code{null.theta.est.ci} in the \code{\link{all_null_theta}} function for the structure of array.}

\item{eflag}{See this argument in the \code{\link{jprat.wrapper}} function.
           If this value is -1, then the marginal distribution \code{F_{es}(t|X, Z)} has missing values (NA).}
}
\description{
gamm4.estimates: used to get ``gamm4" estimate
}
\details{
The estimates depend on whether analysis will be performed separately or jointly. See the argument \code{analyze.separately}
         \code{\link{jprat.main.estimates}}.
}
