% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate.fixed.effects}
\alias{generate.fixed.effects}
\title{Title}
\usage{
generate.fixed.effects(
  num_study,
  nmax,
  np,
  lb.max,
  n,
  lb,
  fzrform,
  par1_fzr,
  par2_fzr,
  type_fzr,
  fxform,
  par1_fx,
  par2_fx,
  type_fx,
  real_data,
  x_tmp.list,
  z_tmp.list
)
}
\arguments{
\item{z_tmp.list}{}
}
\value{

}
\description{
Title
}
