% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.truth}
\alias{get.truth}
\title{Title}
\usage{
get.truth(
  combi.study,
  combi.choice,
  real_data,
  num_study,
  np,
  lb,
  num_time,
  beta0int,
  beta0,
  gamma.param,
  omega.param,
  time_val,
  time_choice.predicted,
  time_choice.predicted.toadd,
  num_xx,
  a0,
  axmod,
  la,
  xks,
  zeval,
  z.choice,
  sigmar,
  sigmau,
  gtmod,
  use.random.effects,
  null.theta,
  combi.null.theta
)
}
\arguments{
\item{combi.null.theta}{}
}
\value{

}
\description{
Title
}
