% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jprat.main.estimates}
\alias{jprat.main.estimates}
\alias{jprat.wrapper}
\alias{gamm4.estimates}
\title{Main Algorithm for JPRAT Estimation}
\usage{
jprat.main.estimates(
  method = "gamm4",
  compute.study.differences = compute.study.differences,
  var.boot = var.boot,
  arbitrary,
  num_study,
  np,
  count.store,
  count.store.outside,
  data,
  num_time,
  time_val,
  time_choice.predicted,
  time_choice.predicted.toadd,
  n,
  nmax,
  m,
  maxm,
  la,
  lb,
  xks,
  truth,
  num_xx,
  knot.length,
  real_data,
  family.data,
  zeval,
  z.choice,
  param.label,
  beta0int,
  gamma.param,
  omega.param,
  spline.constrain,
  common.param.estimation,
  est.cens.depend.z,
  par_fu,
  analyze.separately,
  link.type,
  null.theta,
  z.proportions = NULL,
  betaest.store,
  alphasest.store,
  Ftest.store,
  Ftest.predicted.store,
  combi.study,
  combi.choice,
  combi.names,
  boot = boot,
  boot.null.theta = boot.null.theta,
  boot.combi.null.theta = boot.combi.null.theta,
  betabootci.store = betabootci.store,
  alphasbootci.store = alphasbootci.store,
  Ftbootci.store = Ftbootci.store,
  betadiff.store = betadiff.store,
  alphasdiff.store = alphasdiff.store,
  Ftdiff.store = Ftdiff.store,
  iters = iters
)
}
\arguments{
\item{method}{A character value for which estimation procedure will be used to estimate all parameters of the additive logistic mixed effects model: "gamm4" or "new".
The default option is "gamm4".}

\item{compute.study.differences}{A logical value whether study differences will be estimated: it is only valid when the number of studies (\code{number.of.studies}) are greater than 1,
when parameters were estimated differently (\code{estimated.parameters.common.for.all.studies=FALSE}), and when estimates are similar across studies (\code{check.study.equality=TRUE}).}

\item{var.boot}{A logical value whether bootstrap variances are estimated. See the argument \code{use.bootstrap.variance} in the \code{\link{jprat.wrapper}} function. The default option is TRUE.}

\item{arbitrary}{ADD DETAILS HERE!}

\item{num_study}{The number of studies used in analyses. If the real data analysis used three studies called "cohort", "predict", "pharos", then number of studies is 3. i.e., \code{num_study=3}.}

\item{np}{The number of clinical events, which is the length of the character vector for names of time-to-event outcomes (\code{event.outcome.names}) in \code{\link{jprat.wrapper}} function.}

\item{count.store}{A null object to store number of event times, which occurs before \eqn{t_0}.}

\item{count.store.outside}{A null object to store number of event times, which occurs before \eqn{t_0} for outside [0,1].}

\item{data}{A data set as starting values used for estimation procedure;
a list of starting values for binary event status (y_start), the missing indicator of the binary event status (ymiss_ind_start), the nonfunctional covariate Z (z_start),
the functional covariate X (x_start), the time-to-events or censoring times (s_start), the mixture probabilities for the jack-knife pseudo-values (q_start),
the censoring indicators (delta_start), the original onset ages (onset_age_orig_start), the indicator of no risk (norisk_ind_start), the counting numbers for the status of each event of interests (count),
the pseudo-values (ytest).}

\item{num_time}{A length of the vector of time points at which the predicted values of marginal distributions will be estimated.
See the argument \code{time.points.for.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{time_val}{See the argument \code{time.points.for.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{time_choice.predicted}{See the argument \code{time.points.for.conditional.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{time_choice.predicted.toadd}{See the argument \code{time.points.for.conditional.prediction.toadd} in the \code{\link{jprat.wrapper}} function.}

\item{n}{A vector of the sample sizes for all studies.}

\item{nmax}{The maximum sample size across all studies.}

\item{m}{An array of the number of clinical events for each individual and per study,
whose dimension is the number of studies (\code{num_study})
by the maximum number of the sample sizes across studies (\code{nmax}).}

\item{maxm}{The maximum value of  the events across all subjects for all studies.}

\item{la}{A length of smooth functional parameters \eqn{\alpha(X,t)} in the model. The default value is 1 in the data analysis study.}

\item{lb}{A list of the number (dimension) for the coefficients of nonfunctional covariate Z (see the argument \code{functional.beta.coefficients}) per study.}

\item{xks}{A list of studies whose element of the list is the numeric vector of values for the functional covariates \eqn{X} in smooth functional parameter \eqn{\alpha(X, t)}.}

\item{truth}{A list of true model coefficients for "beta" (\eqn{\beta_{0}(t), \beta_{es}(t)}), "alphas" \eqn{\alpha(X,t)}, "Ft" (\eqn{F_{es}(t|X,Z)}), "Ft.predicted", "beta.diff", and "Ft.diff".}

\item{num_xx}{The number of functional covariate values \eqn{X} at which the functional parameters \eqn{\alpha(X, t)} are evaluated.}

\item{knot.length}{The number of knots used to construct the basis functions.}

\item{real_data}{A logical value whether the real data will be used. The default option is TRUE.}

\item{family.data}{A logical value if the clusters are families, where each member in the cluster is different.}

\item{zeval}{An array of the values of the nonfunctional covariate Z for all studies,
where the order of dimensions for this array is the number of studies (\code{num_study}),
the number of nonfunctional covariate \eqn{Z} (\code{z.choice}),
the length of the character vector for the names of time-to-event outcomes (\code{event.outcome.names})
and the number of coefficients of nonfunctional covariate Z (\code{functional.beta.coefficients}) per study.}

\item{z.choice}{The number of nonfunctional covariate \eqn{Z} used in the analysis.}

\item{param.label}{A character vector of the names for all coefficients of nonfunctional covariate Z in the model.}

\item{beta0int}{A fixed number for the intercept \eqn{\beta_0(t)} in the model. The default value is 0.25.}

\item{gamma.param}{A numeric vector of event specific coefficient, \eqn{\gamma_{e}(t)}. The default value is NULL.}

\item{omega.param}{A numeric vector of study specific coefficient, \eqn{\omega_s(t)}. The default value is NULL.}

\item{spline.constrain}{A logical value if B-spline is constrained at 0. To keep the design matrix correct, we need to have \eqn{B(0)=0}. The default option is TRUE.}

\item{common.param.estimation}{A logical value whether the model parameters are the same across studies: The default option is FALSE. See the argument \code{estimated.parameters.common.for.all.studies} in the \code{\link{jprat.wrapper}} function.}

\item{est.cens.depend.z}{A logical value whether the estimation process assumes that a censoring distribution depends on nonfunctional covariates \eqn{Z}.
If covariates Z does not follow a binomial distribution, then the default option is FALSE. See the argument \code{estimation.when.censoring.depends.on.z} in the \code{\link{jprat.wrapper}} function.}

\item{par_fu}{A vector of parameters for the normal distribution of random effect for all studies: "mean" and "sd".}

\item{analyze.separately}{A character value to determine whether analysis will be performed separately or jointly: the options are "studyevents" (studies and event), "studies", "events", or "none". See the argument \code{what.analyzed.separately} in the \code{\link{jprat.wrapper}} function.}

\item{link.type}{A character value for the name of the link function for the additive mixed effect model (the generalized linear model): "logit” for proportional odds model or "cloglog" for cox proportional hazards. See the argument \code{glm.link.type} in the \code{\link{jprat.wrapper}} function.}

\item{z.proportions}{The marginal probability of the nonfunctional coavariate \eqn{Z}. The default value is NULL.}

\item{betaest.store}{An array to store the estimated functional parameters \eqn{\beta(t)} (beta) at each time point. See the return value \code{null.theta.simus.est.ciboot}
in the \code{\link{all_null_theta}} function.}

\item{alphasest.store}{An array to store the estimated smooth functional parameter \eqn{\alpha(X,t)} (alphas). See the return value \code{null.theta.simus.est.ciboot} in the \code{\link{all_null_theta}} function for the dimension of this array.}

\item{Ftest.store}{An array to store the estimated marginal distribution \eqn{F_{es}(t)} (Ft). See the return value \code{null.theta.simus.est.ciboot} in the \code{\link{all_null_theta}} function for the dimension of this array.}

\item{Ftest.predicted.store}{An array of values for the predicted monotone marginal distributions \eqn{F_{es}(t|X, Z, t>t_{0})}
beyond time \eqn{t_{0}} (\code{Ftest.predicted}) in the \code{\link{gamm4.estimates}} function
and the return value \code{Ft.predicted.var.boot} in the \code{\link{boot.compare.studies}} function at each iteration.
See the return value \code{null.theta.simus.est.ciboot} in the \code{\link{all_null_theta}} function for the dimensions of the array.}

\item{combi.study}{The total number of distinct studies to be compared.}

\item{combi.choice}{A matrix of all combination of two pairs of studies to be compared (paired in column-wise), whose dimension is 2 by the total number of studies (\code{combi.study}).}

\item{combi.names}{A character vector of names for the pairs of study combinations. For example, cohort-predict, cohort-pharos, predict-pharos.}

\item{boot}{The number of bootstrap iterations. The bootstrap procedure is to do hypothesis testing, which compares functional parameters over the time points among studies.
The default value is 100. See the argument \code{number.of.bootstraps} in the \code{\link{jprat.wrapper}} function.}

\item{boot.null.theta}{A list of "null.theta",  "null.theta.simus", "null.theta.simus.ci", "null.theta.ci",
"null.theta.simus.est.ci", "null.theta.simus.est.ciboot", and ``null.theta.est.ci", which are return values in the \code{\link{all_null_theta}} function for the bootstrap estimates.
 The label for simulation dimensions is "boot".}

\item{boot.combi.null.theta}{A list of "null.theta",  null.theta.simus", "null.theta.simus.ci", "null.theta.ci", "null.theta.simus.est.ci",
"null.theta.simus.est.ciboot", and "null.theta.est.ci", which are return values in the \code{\link{all_null_theta}} function
 for comparing bootstrap estimates ("beta”, "alphas”, "Ft” and "Ft.predicted”) among different studies.
 The label for simulation dimensions is "boot".}

\item{betabootci.store}{A multi-dimensional zero array to store the bootstrap estimated functional parameters \eqn{\beta(t)} for comparing study differences.}

\item{alphasbootci.store}{A multi-dimensional zero array to store the bootstrap estimated smooth functional parameters \eqn{\alpha(X, t)} for comparing study differences.}

\item{Ftbootci.store}{A multi-dimensional zero array to store the bootstrap estimated marginal distributions \eqn{Ft_{es}(t|X, Z)} for comparing study differences.}

\item{betadiff.store}{A multi-dimensional zero array to store the differences of the functional parameters \eqn{\beta(t)}
between a pair of studies (cohort-predict, cohort-pharos, predict-pharos) at each time point (\code{time_val}) for each clinical event per study.}

\item{alphasdiff.store}{A multi-dimensional zero array to store the differences of the smooth functional parameters \eqn{\alpha(X, t)}
between a pair of studies (cohort-predict, cohort-pharos, predict-pharos) at each time point (\code{time_val}) for each clinical event per study.}

\item{Ftdiff.store}{A multi-dimensional zero array to store the true values for the differences of the monotone marginal distribution \eqn{F_{es}(t|X,Z)}  between a pair of studies (cohort-predict, cohort-pharos, predict-pharos) at each time point (\code{time_val}) and different covariate values of \eqn{X} and \eqn{Z} for each clinical event per study.}

\item{iters}{The number of iterations for while loops.}
}
\value{
A list of

\item{eflag}{An integer number to check if any error comes up while the JPRAT algorithm is processed.
      This value is 0 if there are no errors. If this value is -1, then the marginal distribution of \eqn{F_{es}(t|X, Z)} has missing values (NA).}
\item{iters}{The number of iterations for while loops.}
\item{count.store}{A data frame for the rate of event times for the uncensored, censored, uncensored but zero, and other cases,
        which depend on the binary status of the events for each subject: "zero" (censored), "one" (uncensored), or "others" (missing).
        The event times are counted at all time points (\code{time_val}) for all studies.}
 \item{count.store.outside}{A data frame of the rate of event times for the uncensored and the censored,
                                which depend on the binary status of the events for each subject outside [0,1]:
                                 "zero" (censored), "one" (uncensored).  The event times are counted at all time points (\code{time_val})
                                 for all studies.}
\item{betaest.store}{An array of values for \eqn{\hat\beta(t)} at each iteration. (See the return value \code{betaest} in the
      \code{\link{gamm4.estimates}} function and the return value \code{beta.var.boot} in  the \code{\link{boot.compare.studies}} function.)}
\item{alphasest.store}{An array of values for the estimated smooth functional parameters \eqn{\alpha(X, t)}
     (See the return value \code{alphasest} in the \code{\link{gamm4.estimates}} function and
      the return value \code{alphas.var.boot} in the \code{\link{boot.compare.studies}} function) at each iteration.
      }
\item{Ftest.store}{An array of values for the estimated monotone marginal distributions \eqn{F_{es}(t|X, Z)} at each iteration.
                  (See the return value \code{Ftest} in the \code{\link{gamm4.estimates}} function
                  and the return value \code{Ft.var.boot} in the  \code{\link{boot.compare.studies}} function.)}
\item{Ftest.predicted.store}{An array of values for the predicted monotone marginal distributions
                            \eqn{F_{es}(t|X, Z, t>t_{0})} beyond time \eqn{t_0} at each iteration.
                            (See the return value \code{Ftest.predicted} in the \code{\link{gamm4.estimates}} function
                            and the return value \code{Ft.predicted.var.boot} in the \code{\link{boot.compare.studies}} function.)}

\item{betabootci.store}{An array of the estimated functional parameters \eqn{\beta(t)} for the study differences
including their estimates (\code{betadiff.store}) and the bootstrap estimates at each iteration (See the return value \code{betabootci}
in the \code{\link{boot.compare.studies}} function).}
\item{alphasbootci.store}{An array of the estimated smooth functional parameters \eqn{\alpha(X, t)} for the study differences
                       including their estimates (\code{alphasdiff.store}) and the bootstrap estimates (see the return value \code{alphasbootci}
                       in the \code{\link{boot.compare.studies}} function) at each iteration.}
\item{Ftbootci.store}{An array of the estimated monotone marginal distribution \eqn{F_{es}(t|X,Z)}
                     for the study differences including their estimates (\code{Ftdiff.store})
                     and the bootstrap estimates (see return value \code{Ftbootci} in the \code{\link{boot.compare.studies}} function)
                     at each iteration. }
}
\description{
This function returns the estimated JPRAT parameters in an additive logistic mixed effect model.
This function also returns a bootstrap estimation procedure to do a hypothesis testing,
which assesses if functional terms between studies differ over a range of time points.
It uses the maximizing double penalized quasi-likelihood (DPQL) to estimate the JPRAT parameters and coefficients
and the generalized approximate cross-validation for the smoothing parameters \eqn{\lambda}.
This can be done using \code{\link[mgcv]{gamm4}} in \code{\link{mgcv}} package in R.
}
\details{
For the structure of the arrays of \code{betabootci.store}, \code{alphasbootci.store} and \code{Ftbootci.store}, see the return value of \code{null.theta.simus.est.ci}
          in the \code{\link{all_null_theta}} function.
         For the structure of the arrays of \code{betaest.store}, \code{alphasest.store},\code{Ftest.store}, and \code{Ftest.predicted.store},
         see the return value of \code{null.theta.simus.est.ciboot} in the \code{\link{all_null_theta}} function.
}
