% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simu.data.fixed.effects}
\alias{simu.data.fixed.effects}
\title{Title}
\usage{
simu.data.fixed.effects(
  x,
  z,
  delta_tmp.list,
  s_tmp.list,
  a0,
  axmod,
  num_time,
  censorrate,
  frform,
  type_fr,
  par1_fr,
  par2_fr,
  par1_fr2,
  par2_fr2,
  mix_n,
  par_fu,
  real_data,
  time_val,
  beta0int,
  beta,
  gamma.param,
  omega.param,
  n,
  nmax,
  m,
  maxm,
  la,
  lb,
  num_study,
  np,
  gtmod,
  use.random.effects,
  gen.cens.depend.z
)
}
\arguments{
\item{gen.cens.depend.z}{}
}
\value{

}
\description{
Title
}
