% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDData_Cohort_Predict_Pharos.R
\docType{data}
\name{simu_data2_model_A}
\alias{simu_data2_model_A}
\title{Generated data set for the Huntington's disease (HD) study:Study2(n=500).}
\format{
A dataframe including 346 subjects' longitudinal data with the following 12 variables:
\describe{
  \item{CAG}{CAG repeats length.}
  \item{base_age}{ages at baseline, when subjects visited clinics for the first time.}
  \item{z2}{ages when HD diagnosis has occurred for subjects who did not have HD diagnosis at the baseline.}
  \item{event1}{ages when the first multi-domain cognitive impairments has occurred for subjects.}
  \item{event2}{ages when depression were occurred for subjects.}
  \item{delta.event1}{censoring indicator for individuals who entered the study without a diagnosis (1 if a person had a HD diagnosis, 0 otherwise).}
  \item{delta.event2}{censoring indicator whether individuals have experienced the first cognitive impairments (1 if a person has experienced cognitive impairment at first, 0 otherwise).}

}
}
\source{
see data-raw/data_generation.R
}
\usage{
simu_data2_model_A
}
\description{
Generated data set for the Huntington's disease (HD) study:Study2(n=500).
}
\keyword{datasets}
