% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{all_null_theta}
\alias{all_null_theta}
\title{all_null_theta: This function is to create a list of all components to be estimated,
where each element of the list is an array with zeros to store all estimates.}
\usage{
all_null_theta(
  theta.names,
  study.names,
  event.names,
  z_lab_names,
  x_lab_names,
  label.dim.simus,
  label.name.simus,
  time_val,
  param.label,
  time_choice.predicted,
  time_choice.predicted.toadd,
  la
)
}
\arguments{
\item{theta.names}{a character vector of names that will be estimated in the model: functional parameters
\eqn{\beta_0(t)} , \eqn{\beta_{es}(t)}, ("beta") \eqn{\alpha(X, t)},("alphas") and \eqn{F_{es}(t)} ("Ft" and "Ft.predicted").}

\item{study.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{event.names}{See the argument \code{event.outcome.names} in the \code{\link{jprat.wrapper}} function.}

\item{z_lab_names}{a vector of character values for the names of nonfunctional covariates Z.}

\item{x_lab_names}{a vector of character values for the names of functional covariates X in functional coefficients \eqn{\alpha(X, t)}.}

\item{label.dim.simus}{number of simulation (bootstrap iterations) runs: "simus" for simulation runs or "boot" for bootstrap iterations.}

\item{label.name.simus}{a vector of names to label dimensions to be added to an exsisting array. The length of this vector is the number of simulation or bootstrap iterations (label.dim.simus).}

\item{time_val}{a vector of time points at which the predicted values of marginal distribution marginal probabilities of outcomes will be estimated.
See the argument \code{time.points.for.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{param.label}{a character vector of the names for all coefficients of nonfunctional covariate Z in the model.}

\item{time_choice.predicted}{See the argument \code{time.points.for.conditional.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{time_choice.predicted.toadd}{See the argument \code{time.points.for.conditional.prediction.toadd} in the \code{\link{jprat.wrapper}} function.}

\item{la}{length of smooth functional parameters \eqn{\alpha(X,t)} in the model. In real analysis, we only have one  \eqn{\alpha(X,t)}. Default is 1.}
}
\value{
A list of

\item{null.theta}{a list of all components to be estimated,
                  whose elements are "beta", "alphas", "Ft" and ``Ft.predicted".
                  Each of element of the list is an array of zeros with a specific dimensions.}
\item{null.theta.simus}{a list of all components to be estimated (null.theta),
                       where each array of the list has extended values
                       for the number of simulation (label.dim.simus)
                       added to the first dimension of the array.}
\item{null.theta.ci}{a list of all components to be estimated (null.theta),
                   where each array of the list has extended values for the estimated variance,
                   the lower bound of confidence interval, and the upper bound of confidence interval,
                   which are added to the last dimension of the array.}
\item{null.theta.est.ci}{a list of all components to be estimated (null.theta),
                          where each array of the list has extended values for the estimates,
                          estimated variance, the lower bound of confidence interval,
                          and the upper bound of confidence interval,
                          which are added to the last dimension of the array.}
\item{null.theta.simus.ci}{a list of all components to be estimated (null.theta),
                          where the first dimension of each array is the number of simulations (null.theta.simus),
                          and where each array of the list has extended values for the estimated variance,
                          the lower bound of confidence interval,
                          and the upper bound of confidence interval,
                          which are added to the last dimension of the array.}
\item{null.theta.simus.est.ci}{a list of all components to be estimated (null.theta),
                             where the first dimension of each array is the number of simulations (null.theta.simus),
                             and where each array of the list has extended values for the estimates, the estimated variance,
                             the lower bound of confidence interval, and the upper bound of confidence interval,
                             which are added to the last dimension of the array.}
\item{null.theta.simus.est.ciboot}{a list of all components to be estimated (null.theta), where the first dimension of each array is the number of simulations (null.theta.simus), and where each array of the list has extended values for the estimates,
                                  the estimated variance, the lower bound of confidence interval,
                                  the upper bound of confidence interval,
                                  the estimated bootstrap variances,
                                  the estimated lower bound of bootstrap confidence interval,
                                  and the estimated upper bound of bootstrap confidence interval,
                                   which are added to the last dimension of the array.}
}
\description{
all_null_theta: This function is to create a list of all components to be estimated,
where each element of the list is an array with zeros to store all estimates.
}
\details{
The list will be created to store estimates for "beta", "alphas", "Ft" and "Ft.predicted".
}
