% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jprat_function.R
\name{all_null_theta}
\alias{all_null_theta}
\title{Initialization of parameters and coefficients}
\usage{
all_null_theta(
  theta.names,
  study.names,
  event.names,
  z_lab_names,
  x_lab_names,
  label.dim.simus,
  label.name.simus,
  time_val,
  param.label,
  time_choice.predicted,
  time_choice.predicted.toadd,
  la
)
}
\arguments{
\item{theta.names}{A character vector of functional parameters' names that will be estimated in the model: "beta" (include intercept), "alphas" and "Ft" (or "Ft.predicted") corresponds to
\eqn{\beta_{es}(t)} (\eqn{\beta_0(t)}) ,  \eqn{\alpha(X, t)}, and \eqn{F_{es}(t)}.}

\item{study.names}{a character vector of name of studies used in analyses. e.g., c("cohort", "predict", "pharos").}

\item{event.names}{See the argument \code{event.outcome.names} in the \code{\link{jprat.wrapper}} function.}

\item{z_lab_names}{A vector of character values for the names of nonfunctional covariates Z.}

\item{x_lab_names}{A vector of character values for the names of functional covariates X in the functional coefficients \eqn{\alpha(X, t)}.}

\item{label.dim.simus}{The number of simulation runs (or bootstrap iteration runs), which determines the length of the dimension when extra dimensions are added to store the simulation results.}

\item{label.name.simus}{A character string of names that will be used to label the dimensions (in row wise) whose lengths are given by \code{label.dim.simus} when extra dimensions are added to store the simulation results.}

\item{time_val}{See the argument \code{time.points.for.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{param.label}{A character vector of the names for all coefficients of nonfunctional covariate Z in the model.}

\item{time_choice.predicted}{See the argument \code{time.points.for.conditional.prediction} in the \code{\link{jprat.wrapper}} function.}

\item{time_choice.predicted.toadd}{See the argument \code{time.points.for.conditional.prediction.toadd} in the \code{\link{jprat.wrapper}} function.}

\item{la}{A length of smooth functional parameters \eqn{\alpha(X,t)} in the model. The default value is 1 in the data analysis study.}
}
\value{
A list of

\item{null.theta}{A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted") in the model.
                 Each element of the list contains arrays of zeros at the specific time points, the values of covariates and the event of interest in each study,
                 whose dimension determined by the length of studies, events, time points
                 where prediction occurs, and coefficients of the nonfunctional covariates.}


\item{null.theta.simus}{A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted"), where each element of the list contains arrays of zeros.
 Each element has an additional dimension as the first dimension of each array, which is extended arrays from \code{null.theta}
 The number of simulations determines the length of the additional dimension (\code{label.dim.simus}) with its dimension names (\code{label.names.simus}) as a character string.}


\item{null.theta.ci}{A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted"), where each element of the list contains arrays of zeros.
, where three dimensions for the estimated variance, the lower and upper bounds of the confidence intervals were added to the last dimension of \code{null.theta}.}


\item{null.theta.est.ci}{A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted"), where each element of the list contains arrays of zeros.
Four dimensions for the estimates, the estimated variance, the lower and upper bounds of the confidence intervals were added to the last dimension of \code{null.theta}.}


\item{null.theta.simus.ci}{
A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted"),
where each element of the list contains arrays of zeros. Three dimensions for the estimated variance, the lower and upper bounds of the confidence intervals
were added to the last dimension of \code{null.theta.simus}.}

\item{null.theta.simus.est.ci}{A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted"),
where each element of the list contains arrays of zeros. Four dimensions for the estimates, the estimated variance, the lower and upper bounds of the confidence intervals
were added to the last dimension of \code{null.theta.simus}.}


\item{null.theta.simus.est.ciboot}{A list of all estimated parameters and coefficients ("beta", "alphas", "Ft" and ``Ft.predicted"),
where each element of the list contains arrays of zeros. Seven dimensions for the estimates, the estimated variance, the lower and upper bounds of the confidence intervals,
the estimated bootstrap variances, the lower and upper bounds of bootstrap confidence intervals were added to the last dimension of \code{null.theta.simus}.}
}
\description{
This function is to initialize a list of all components to be estimated to zeros. The structure of each component is a multi-dimensional array, whose dimension size depends on the number of studies, the number of events,
the number of covariates, and the length of time points where estimates will be evaluated, etc.
}
\details{
A list of zero arrays will be created to store those estimates of \eqn{\beta_{es}(t)} (\eqn{\beta_0(t)}) ,  \eqn{\alpha(X, t)}, and \eqn{F_{es}(t)} in the time varying, proportional odds model with random effects.
}
\references{
Garcia, T.P., Marder, K., and Wang, Y. (2019). Time-varying proportional odds model for mega-analysis of clustered event times. Biostatistics, 20(1), 129-146.
}
