% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{boot.compare.studies}
\alias{boot.compare.studies}
\title{boot.compare.studies: This function estimates the bootstrap estimates for study comparison.}
\usage{
boot.compare.studies(
  arbitrary,
  combi.study,
  combi.choice,
  combi.names,
  num_study,
  boot,
  np,
  data,
  num_xx,
  num_time,
  time_val,
  time_choice.predicted,
  time_choice.predicted.toadd,
  xks,
  n,
  nmax,
  m,
  maxm,
  la,
  lb,
  truth,
  real_data,
  knot.length,
  family.data,
  zeval,
  z.choice,
  param.label,
  beta0int,
  gamma.param,
  omega.param,
  spline.constrain,
  common.param.estimation,
  est.cens.depend.z,
  par_fu,
  analyze.separately,
  link.type,
  boot.null.theta,
  boot.combi.null.theta,
  var.boot,
  compute.study.differences,
  est.values,
  z.proportions
)
}
\arguments{
\item{arbitrary}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{combi.study}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{combi.choice}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{combi.names}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{num_study}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{boot}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{np}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{data}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{num_xx}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{num_time}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{time_val}{See the argument in the \code{\link{all_null_theta}} function.}

\item{time_choice.predicted}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{time_choice.predicted.toadd}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{xks}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{n}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{nmax}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{m}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{maxm}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{la}{See the argument in the \code{\link{all_null_theta}} function.}

\item{lb}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{truth}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{real_data}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{knot.length}{number of knots used to construct the basis functions.}

\item{family.data}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{zeval}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{z.choice}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{param.label}{See the argument in the \code{\link{all_null_theta}} function.}

\item{beta0int}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{gamma.param}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{omega.param}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{spline.constrain}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{common.param.estimation}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{est.cens.depend.z}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{par_fu}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{analyze.separately}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{link.type}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{boot.null.theta}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{boot.combi.null.theta}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{var.boot}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{compute.study.differences}{See the argument in the \code{\link{jprat.main.estimates}} function.}

\item{est.values}{a list of arrays of the estimates ("est") for the functional coefficient \eqn{\beta(t)}
(\code{betaest} in the \code{\link{gamm4.estimates}} function),
the smooth functional coefficient \eqn{\alpha(X,t)} (\code{alphas} in the \code{\link{gamm4.estimates}} function),
the monotone marginal distribution \eqn{F_{es}(t|X, Z)} (\code{Ft} in the \code{\link{gamm4.estimates}}) function
and the predicted values of the monotonic marginal distribution \code{F_{es}(t|X, Z, t > t0)}
(\code{Ft.predicted} in the \code{\link{gamm4.estimates}} function).}

\item{z.proportions}{See the argument in the \code{\link{jprat.main.estimates}} function.}
}
\value{
A list of

\item{betabootci}{array of the differences of estimated bootstrap functional parameters \eqn{\beta(t)}
                  between a pair of studies at each time point for each clinical event,
                  which includes values of the estimated bootstrap variances,
                  the lower bound of the bootsrap confidence interval,
                  and the upper bound of confidence intervals for \eqn{\beta(t)}.
                  For the dimension, see the argument \code{null.theta.ci} in the \code{[\link{all_null_theta}}.}
\item{alphasbootci}{array of the differences of estimated bootstrap smooth functional parameters \eqn{\alpha(X, t)}
                    between a pair of studies at each time point and the functional covariate value X for each clinical event.
                    For the dimension, see the argument \code{null.theta.ci} in the \code{\link{all_null_theta}}.}
\item{Ftbootci}{array of the differences of estimated bootstrap marginal distribution \eqn{Ft_{es}(t|X, Z)}
                between a pair of studies at each time point, each functional covariate value X and
                each nonfunctional  covariate value Z for each clinical event.
                For the dimension, see the argument \code{null.theta.ci} in the\code{\link{all_null_theta}}.}
\item{beta.var.boot}{array of the estimated bootstrap functional parameters \eqn{\beta(t)} at each time point
                     and each clinical event per study, which includes values of the estimated bootstrap variances,
                     the lower bound of the bootsrap confidence interval, and the upper bound of confidence intervals for \eqn{\beta(t)}.
                     For the dimension, see the argument \code{null.theta.ci} in the \code{\link{all_null_theta}}.}
\item{alphas.var.boot}{array of the estimated bootstrap smooth functional parameters \eqn{\alpha(X, t)} at each time point
                       and each functional covariate value X for each clinical event per study.
                       For the dimension, see the argument \code{null.theta.ci} in the \code{\link{all_null_theta}}.}
\item{Ft.var.boot}{array of the estimated bootstrap marginal distribution \eqn{Ft_{es}(t|X, Z)}
                   at each time point, each functional covariate value X
                   and each nonfunctional covariate value Z for each clinical event per study.
                   For the dimension, see the argument \code{null.theta.ci} in the \code{\link{all_null_theta}}.}
\item{Ft.predicted.var.boot}{array of the estimated predcited bootstrap marginal distribution \eqn{Ft_{es}(t|X, Z, t>t_{0})}
                            at each time \eqn{t},
                            each functional covariate value X and each nonfunctional covariate value Z for each clinical event per study
                            given that a subject does not experience each event by \eqn{t_0}.
                            For the dimension, see the argument \code{null.theta.ci} in the \code{\link{all_null_theta}}.}
\item{count.store.boot}{a data frame for the bootstrap rate of event times for uncensored, censored, uncensored but zero, and other cases,
                        which depends on the binary status of the event for each subject: "zero" (censored), "one" (uncensored), or "others" (missing).
                        See "count.store" in the \code{\link{gamm4.estimates}} for details.}
\item{count.store.outside.boot}{a data frame of the bootstrap rate of event times for uncensored and censored,
                                which depends on the binary status of the event for each subject,
                                which are outside [0,1]: "zero" (censored), "one" (uncensored).
                                See the argument "count.store.outside" in the \code{\link{gamm4.estimates}} for details.}
}
\description{
boot.compare.studies: This function estimates the bootstrap estimates for study comparison.
}
\details{
This function uses bootstrap-based joint confidence interval to determine which model
         (joint model with distinct study parameters, shared study parameter model, separate study model)
         is more suitable.
}
