% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.number.at.risk.for.HD}
\alias{compute.number.at.risk.for.HD}
\title{Number at risk for HD data}
\usage{
compute.number.at.risk.for.HD(
  study.names,
  input.data.list,
  event.outcome.names,
  nonfunctional.covariate.names,
  functional.covariate.names,
  nonfunctional.covariate.value,
  functional.covariate.values.of.interest,
  time.points.for.prediction,
  estimated.parameters.common.for.all.studies,
  write.output
)
}
\arguments{
\item{study.names}{A character vector of name of studies used in the analysis. e.g., c("cohort", "predict", "pharos").}

\item{input.data.list}{A list of datasets to be analyzed. In the case, you use datasets given from JPRAT, then there are two options:
use list(cohort=data_cohort, predict=data_predict, pharos=data_pharos) if you choose \code{what.analyzed.separately="event" or "eventstudy"};
otherwise use list(study1=simu_data1_model_A, study2=simu_data2_model_A) (See \code{what.analyzed.separately="none" or "study"} for option details).}

\item{event.outcome.names}{A character vector for the name of the time-to-event of interest (outcomes).}

\item{nonfunctional.covariate.names}{A character vector for the name of the nonfunctional covariates that is used in the analysis.}

\item{functional.covariate.names}{A character vector for the name of the functional covariates that is used in the analysis.}

\item{nonfunctional.covariate.value}{A numeric value interested for the nonfunctional covariate. e.g., 40 for \code{base_age}, which will be used for prediction.}

\item{functional.covariate.values.of.interest}{A vector of specific functional covariate values \eqn{X=x} where the smooth functional parameter \eqn{\alpha(X=x,t)} will be estimated. For example, we set \eqn{x=46,48,50} in the analysis. We recommend that the specific functional covariate values \eqn{X=x} should be chosen between the minimum and the maximum of functional covariate values. Otherwise, users may encounter a warning message.}

\item{time.points.for.prediction}{A vector of time points at which the predicted values of marginal distribution will be estimated. Users may encounter a warning message that "Not enough data at time points (e.g., base_age);  May have divergent in the integration while bootstrapping." We recommend that users evaluate the proportional odds model at the average of nonfunctional covariate (e.g., base_age) across all subjects in each study. This may get rid of the warning messages.}

\item{estimated.parameters.common.for.all.studies}{A logical value whether the model parameters are same across studies. The default option is FALSE.}

\item{write.output}{A logical value whether outputs should be saved into data files. The default option is TRUE.}
}
\value{
a table for the number of subjects who are at risk for HD for each study at times.
}
\description{
This function returns the number of subjects who are at risk for HD for each study at times.
}
