% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert.new.notation.to.old.for.get.results}
\alias{convert.new.notation.to.old.for.get.results}
\title{convert.new.notation.to.old.for.get.results: This code converts the new notation into the original code notation for getting results (plots)}
\usage{
convert.new.notation.to.old.for.get.results(
  study.names,
  time.points.for.prediction,
  time.points.for.conditional.prediction,
  time.points.for.conditional.prediction.toadd,
  nonfunctional.covariate.names,
  nonfunctional.covariate.values.for.prediction,
  functional.covariate.names,
  functional.covariate.values.of.interest,
  functional.covariate.values.of.interest.ci,
  functional.covariate.values.for.prediction,
  xmin,
  xmax,
  event.outcome.names,
  use_real_data,
  use.functional.beta.intercept,
  use.functional.event.coefficients,
  use.functional.study.coefficients,
  number.of.bootstraps,
  what.analyzed.separately,
  estimated.parameters.common.for.all.studies,
  glm.link.type,
  use.bootstrap.variance,
  time.points.of.interest,
  time.points.of.interest.ci,
  label.for.alpha.values.over.time,
  nonfunctional.covariate.comparisons,
  add.number.at.risk.legend,
  ylabel.for.plots.comparing.studies,
  xlabel.for.plots.comparing.studies,
  plot.confidence.intervals,
  color.labels,
  legend.labels,
  event.comparison.table,
  functional.covariate.comparisons,
  functional.covariate.comparisons.for.sample.size
)
}
\description{
convert.new.notation.to.old.for.get.results: This code converts the new notation into the original code notation for getting results (plots)
}
