% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data.reformatted.for.analysis.results}
\alias{data.reformatted.for.analysis.results}
\title{data.reformatted.for.analysis.results: This code reform datasets for JPRAT to be able to understand the given datasets to produce results tables and plots.}
\usage{
data.reformatted.for.analysis.results(
  study.names,
  event.outcome.names = NULL,
  color.names,
  legend.names,
  which.nonfunctional.covariate.comparisons = NULL
)
}
\arguments{
\item{study.names}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{event.outcome.names}{See the argument in the \code{\link{jprat.wrapper}} function.}

\item{color.names}{See the argument in the \code{\link{view.all.results}} function.}

\item{legend.names}{See the argument in the \code{\link{view.all.results}} function.}

\item{which.nonfunctional.covariate.comparisons}{See the argument in the \code{\link{view.all.results}} function.}
}
\value{

}
\description{
data.reformatted.for.analysis.results: This code reform datasets for JPRAT to be able to understand the given datasets to produce results tables and plots.
}
\examples{

}
