% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.cis}
\alias{get.cis}
\title{get.cis: This function generates a data frame, which contains the confidence intervals for the estimates,
whether the signs of the lower bounds and upper bounds of the confidence intervals will be changed,
and the length of confidence intervals at each time for specific covariate values for all studies.}
\usage{
get.cis(
  out,
  flatten.name,
  theta = "alpha",
  time_choice,
  xx_val,
  xmin,
  xmax,
  xx_choose,
  zz_choose = NULL,
  convert,
  round.set,
  var.lo,
  var.hi,
  est,
  noshow = NULL,
  track.sign.change = TRUE
)
}
\arguments{
\item{out}{array of estimates to plot.}

\item{flatten.name}{a character value of the flatten name.}

\item{theta}{a character value for name of the component estimated: "beta" (coefficients of the nonfunctional covariates Z),
"alphas" (coefficients of the functional covariates X),  "F" (monotone marginal distributions \eqn{F_{es}(t|Z, X)})
 or "Ft.pred" (the predicted marginal distribution \eqn{F_{es}(t|Z, X, t>t_{0})}.
 Null for the parameter "beta".}

\item{time_choice}{a vector of time points at which confidence intervals of estimates will be predicted over those times. See the argument \code{time.points.of.interest.ci} in the \code{\link{view.all.results}}.}

\item{xx_val}{a vector of values for functional covariates X per a study. See the argument \code{functional.covariate.values.for.prediction} in the \code{\link{view.all.results}} function.}

\item{xmin}{minimum value for the functional covariate values of X.}

\item{xmax}{maximum value for the functional covariate values of X.}

\item{xx_choose}{a vector of specific functional covariate values of X
at which confidence interval of the smooth functional parameter \eqn{\alpha(x,t)} will be predicted.
Null for the parameter "beta".}

\item{zz_choose}{a vector of character values for the names of nonfunctional covariates Z.
For example, zz_choose=z_lab_names. Null for the parameter "beta" and "alphas".}

\item{convert}{a logical value whether the CAG repeat lengths will be scaled into the uniformly distributed values on [0,1].}

\item{round.set}{number of significant digits to use in the plot: Default is 2.}

\item{var.lo}{a character value for the name of the lower bound for the confidence interval: "varlo".}

\item{var.hi}{a character value for the name of the upper bound for the confidence interval: "varhi".}

\item{est}{a character value for the estimates, e.g., "est".}

\item{noshow}{a character vector of names, which do not show in the data frame.}

\item{track.sign.change}{a logical value whether the sign change will be tracking. Default is TRUE.}
}
\value{
\item{out.flatten}{a data frame, created from the array "out" using the \code{flatten.array} function.
                   The data frame contains the confidence intervals for the estimates (ci),
                   whether the signs of the lower bounds and upper bounds of the confidence intervals for the estimates will be changed (\code{sign.change}),
                   and the length of confidence intervals (\code{ci.length}) at specific time points (\code{time_choice}),
                   covarites values (\code{xx_choose}, \code{zz_choose}) and the clinical events of interest for all studies.}
}
\description{
get.cis: This function generates a data frame, which contains the confidence intervals for the estimates,
whether the signs of the lower bounds and upper bounds of the confidence intervals will be changed,
and the length of confidence intervals at each time for specific covariate values for all studies.
}
\details{
The data for "noshow" will not be shown in the data frame.
         If the signs for the lower bounds and upper bounds of the confidence intervals are the same,
         the change of sign (\code{sign.change}) returns TRUE (or 1); otherwise, FALSE (or 0).
         This values will be used to test whether the study results are similar (whether the true model share the study parameters).
}
\examples{


}
