% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggplot_error_bars}
\alias{ggplot_error_bars}
\title{ggplot_error_bars: This function plots the line plots with error bars.}
\usage{
ggplot_error_bars(
  filename,
  estimate,
  theta.array.lo = NULL,
  theta.array.hi = NULL,
  index_a,
  bvalues,
  color.list,
  cex.line = 1.5,
  cex.size = 20,
  cex.number = 6,
  ylim = NULL,
  conf.int = FALSE,
  label.names = NULL,
  ylab.use = "",
  xlab.use = ""
)
}
\arguments{
\item{filename}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{estimate}{array for the estimates (estimated parameters or estimated averaged parameters)
with corresponding time points at each event.}

\item{theta.array.lo}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{theta.array.hi}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{index_a}{a character vector of names for the clinical events.}

\item{bvalues}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{color.list}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{cex.line}{magnification of thickness of the error bar relative to cex. Default is 1.5.}

\item{cex.size}{magnification of size of text in the legend relative to cex. Default is 20.}

\item{cex.number}{magnification of size of text in the labels relative to cex. Default is 6.}

\item{ylim}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{conf.int}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{label.names}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{ylab.use}{See the \code{\link{ggplot_at_a_over_b}} function.}

\item{xlab.use}{See the \code{\link{ggplot_at_a_over_b}} function.}
}
\value{
line plots with errors bars for each event occured at specific time points (\code{bvalues}).
}
\description{
ggplot_error_bars: This function plots the line plots with error bars.
}
