% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{view.all.results}
\alias{view.all.results}
\title{Output of JPRAT}
\usage{
view.all.results(
  study.names,
  input.data.list,
  nonfunctional.covariate.names,
  functional.covariate.names,
  othercovariate.names,
  event.outcome.names,
  delta.names,
  time.points.for.prediction,
  time.points.for.conditional.prediction,
  time.points.for.conditional.prediction.toadd,
  nonfunctional.covariate.value,
  functional.covariate.values.of.interest,
  number.of.bootstraps,
  use.functional.beta.intercept,
  use.functional.event.coefficients,
  use.functional.study.coefficients,
  check.study.equality,
  estimated.parameters.common.for.all.studies,
  what.analyzed.separately,
  estimation.when.censoring.depends.on.z,
  use.bootstrap.variance,
  functional.covariate.values.of.interest.ci,
  time.points.of.interest,
  time.points.of.interest.ci,
  label.for.alpha.values.over.time = NULL,
  color.names,
  legend.names,
  functional.covariate.comparisons,
  functional.covariate.comparisons.for.sample.size = NULL,
  do.plots,
  plot.confidence.intervals,
  add.number.at.risk.legend,
  ylabel.for.plots.comparing.studies,
  xlabel.for.plots.comparing.studies,
  file.name.for.analysis,
  jprat.output = jprat.output,
  nrisk = nrisk,
  show.results.description = show.results.description,
  is.nrisk.data.frame = is.nrisk.data.frame,
  write.jprat.output = write.jprat.output
)
}
\arguments{
\item{study.names}{A character vector of name of studies used in the analysis. e.g., c("cohort", "predict", "pharos").}

\item{input.data.list}{A list of datasets to be analyzed. In the case, you use datasets given from JPRAT, then there are two options:
use list(cohort=data_cohort, predict=data_predict, pharos=data_pharos) if you choose \code{what.analyzed.separately="event" or "eventstudy"};
otherwise use list(study1=simu_data1_model_A, study2=simu_data2_model_A) (See \code{what.analyzed.separately="none" or "study"} for option details).}

\item{nonfunctional.covariate.names}{A character vector for the name of the nonfunctional covariates that is used in the analysis.}

\item{functional.covariate.names}{A character vector for the name of the functional covariates that is used in the analysis.}

\item{othercovariate.names}{A character vector for the name of other covariates, which are not to be analyzed but kept within data during the analysis procedure.}

\item{event.outcome.names}{A character vector for the name of the time-to-event of interest (outcomes).}

\item{delta.names}{A character vector for the name of the censoring indicator for the time-to-event outcomes.}

\item{time.points.for.prediction}{A vector of time points at which the predicted values of marginal distribution will be estimated. Users may encounter a warning message that "Not enough data at time points (e.g., base_age);  May have divergent in the integration while bootstrapping." We recommend that users evaluate the proportional odds model at the average of nonfunctional covariate (e.g., base_age) across all subjects in each study. This may get rid of the warning messages.}

\item{time.points.for.conditional.prediction}{A vector of time points \eqn{t} at which the predicted values of conditional distribution \eqn{Pr(T <t + t_{0}| T>t)} will be estimated.}

\item{time.points.for.conditional.prediction.toadd}{A vector of future time points \eqn{t_{0}} at which the predicted values of conditional distribution \eqn{Pr(T <t + t_{0} | T>t)} will be estimated.}

\item{nonfunctional.covariate.value}{A numeric value interested for the nonfunctional covariate. e.g., 40 for \code{base_age}, which will be used for prediction.}

\item{functional.covariate.values.of.interest}{A vector of specific functional covariate values \eqn{X=x} where the smooth functional parameter \eqn{\alpha(X=x,t)} will be estimated. For example, we set \eqn{x=46,48,50} in the analysis. We recommend that the specific functional covariate values \eqn{X=x} should be chosen between the minimum and the maximum of functional covariate values. Otherwise, users may encounter a warning message.}

\item{number.of.bootstraps}{Number of iterations used in bootstrap estimation.}

\item{use.functional.beta.intercept}{A logical value whether a functional intercept \eqn{\beta_{0}(t)} is included in the time-varying proportional odds model. The default option is TRUE.}

\item{use.functional.event.coefficients}{A logical value whether the event specific effect \eqn{\gamma_{e}(t)} will be used in the time-varying proportional odds model. The default option is TRUE.}

\item{use.functional.study.coefficients}{A logical value whether the study specific effect \eqn{\omega_{s}(t)} will be used in the time-varying proportional odds model. The default option is TRUE.}

\item{check.study.equality}{A logical value whether estimates are similar across studies: The default option is TRUE when what.analyzed.separately is "none"; otherwise, FALSE.}

\item{estimated.parameters.common.for.all.studies}{A logical value whether the model parameters are same across studies. The default option is FALSE.}

\item{what.analyzed.separately}{A character value to determine whether the event information from all studies is analyzed jointly or separately with distinct or shared study parameters in the model: the options are "studyevent" (studies and event), "study", "event", or "none".}

\item{estimation.when.censoring.depends.on.z}{A logical value whether the estimation process assumes that a censoring distribution depends on nonfunctional covariates \eqn{Z}. If covariates \eqn{Z} does not follow a binomial distribution, then the default option is FALSE.}

\item{use.bootstrap.variance}{A logical value whether the bootstrap variances will be evaluated. The default option is TRUE.}

\item{functional.covariate.values.of.interest.ci}{A vector of specific functional covariate values of \eqn{X},
 where the confidence intervals of smooth functional parameter \eqn{\alpha(X=x,t)} will be estimated.
 For example, we set the values \eqn{x=46, 51} at specific times \eqn{t}in the analysis.
 We recommend users to choose the same vector of functional covariate values \eqn{X} as in the interested functional covariate values (\code{functional.covariate.values.of.interest})
to ease estimation procedure in the analysis.}

\item{time.points.of.interest}{A vector of specific time points at which the smooth functional parameters \eqn{\alpha(X,t)}
will be predicted. These time points are used to be labeled for the plot.
For example, we set the values \eqn{x=46, 56} in the analysis.
We recommend users to choose these time points vector as the subset of time points for prediction (\code{time.points.for.prediction}).}

\item{time.points.of.interest.ci}{A vector of time points at which confidence intervals of \eqn{\alpha(X,t)} are predicted.
These time points are used to be labeled for the plot. For example, we set the values \eqn{x=46, 51, 56, 61, 66} in the analysis.
We recommend that these time points vector should be chosen as the subset of time points for prediction (\code{time.points.for.prediction}) or as the same time points for estimating \eqn{\alpha(X,t)} (\code{time.points.of.interest}).}

\item{label.for.alpha.values.over.time}{A vector of specific time points at which the smooth functional parameters \eqn{\alpha(X,t)} will be predicted.
These time points are used to be labeled for the plot. Currently, we do not make a plot for the smooth functional parameters \eqn{\alpha(X,t)}.
The default is NULL.}

\item{color.names}{A character vector to denote which colors will be used for the labels of events of interest in plots. e.g., c("firebrick1", "darkgreen", "black").}

\item{legend.names}{A character vector to label event of interests in plots. e.g., c("Motor Diagnosis (DCL=4)",  "Cognitive Impairment" , "Stage II TFC").}

\item{functional.covariate.comparisons}{A vector of functional covariate values of \eqn{X} at which the smooth functional parameters \eqn{\alpha(X=x,\cdot)} were estimated and compared.
For example, the functional covariate values are \eqn{x=c(46, 51)} and the parameters of \eqn{\alpha(x=46,\cdot)} and \eqn{\alpha(x=51,\cdot)} will be compared.
We recommend that these covariate values \eqn{X} should be the same as the vector of functional covariate values of \eqn{X}
for confidence intervals (\code{functional.covariate.values.of.interest.ci}).}

\item{functional.covariate.comparisons.for.sample.size}{A specific functional covariate values of \eqn{X} to use the sample size calculation,
which will be compared across studies. Currently, we are not interested in sample size calculation. The default is NULL.}

\item{do.plots}{A logical value whether the plot will be generated as results. The default is TRUE.}

\item{plot.confidence.intervals}{A logical value whether confidence intervals of the predicted marginal distributions will be shown in the plot. The default is TRUE.}

\item{add.number.at.risk.legend}{A logical value whether number of subjects who are at risks at the interested time points will be added with legends in plots. The default is TRUE.}

\item{ylabel.for.plots.comparing.studies}{A character value for y-label in plots when studies are comparing or event of interests are compared in each study. The default is "Probability".}

\item{xlabel.for.plots.comparing.studies}{A character value for x-label in plots when studies are comparing for each event of interest or event of interests are compared in each study. The default is "Age(years)".}

\item{file.name.for.analysis}{A character value, which is used as the part of plot names.}

\item{jprat.output}{A list of estimated values. See argument \code{write.ouput} in \code{\link{jprat.wrapper}} function. If \code{write.ouput=TRUE}, the jprat function returns the results as a list of dataframes; otherwise, it returns the results as a list of arrays for the estimated values.
See return values in \code{\link{jprat.wrapper}} function for details.}

\item{nrisk}{Data file for the number of subjects who are at risk at the interested the time points.}

\item{show.results.description}{A logical value whether the description of plots and tables will be produced. The default is FALSE.}

\item{is.nrisk.data.frame}{A logical value whether output for nrisk is the data frame. If write.output=TRUE  (see input argument for \code{\link{compute.number.at.risk.for.HD}} function), it should be TRUE; Otherwise, FALSE.}

\item{write.jprat.output}{A logical value whether output for the jprat function will be created. If write.output=TRUE (see input argument for \code{\link{jprat.wrapper}} function), then it should be TRUE; Otherwise, FALSE.}
}
\value{
This function returns the results of JPRAT algorithm via ggplots and the results whether study parameters were different using the bootstrap based joint confidence interval.
}
\description{
This function returns the results of JPRAT estimates using ggplots and
             the evaluation of study differences between functional parameters
             using the bootstrap-based joint confidence intervals.
}
\details{
We generate two different datasets and apply them to JPRAT algorithm. One is generated by mimicking the real data used in Section 4 (Garcia et al., Biostatistics, 2017) and the other is generated using Model A (Joint model with disticnt study parameters) with parameter setting A.
         The ggplots were shown for:
         \itemize{
         \item Estimated marginal distribution functions with pointwise 95% confidence band for HD motor diagnosis with age at study entry (46 years) and CAG repeat-length (40 repeats).
         \item Predicted probability  (and error bars) an individual with 47 CAG repeats who is 46, 51, 56 experiences motor-diagnosis.
}
}
\examples{
## remove history         ##
## rm(list=ls(all=TRUE))  ##
############################

## Four options whether study and event are analyzed separately:
## what.analyzed.separately="studyevent", "event", "study" or "none"

     what.analyzed.separately= "none"


## Load your data in the list format: should match order of study.names
## Choose different list of data depending on the settings such as
    if(what.analyzed.separately=="event"||what.analyzed.separately=="studyevent"){

       input.data.list=list(cohort=data_cohort, predict=data_predict, pharos=data_pharos);

    }else if(what.analyzed.separately=="none"||what.analyzed.separately=="study"){

       input.data.list=list(study1=simu_data1_model_A, study2=simu_data2_model_A);
    }




###################################################################
##Input data for JPRAT estimation procedure: paramaeter settings ##
###################################################################

    if(what.analyzed.separately=="event"||what.analyzed.separately=="studyevent"){


      study.names=c("cohort", "predict", "pharos");
      othercovariate.names=c("firstyear", "lastyear");
      event.outcome.names=c("hdage_nobase", "mcione", "dep2");
      delta.names=c("delta.hdage_nobase", "delta.mcione", "delta.dep2");
      nonfunctional.covariate.names=c("base_age");
      functional.covariate.names="CAG";
      time.points.for.prediction=seq(46, 66, by=5);
      time.points.for.conditional.prediction=c(46,51, 56);
      time.points.for.conditional.prediction.toadd=c(5);
      nonfunctional.covariate.value=c(40);
      functional.covariate.values.of.interest=c(46, 48, 50) ;

    }else if(what.analyzed.separately== "none"||what.analyzed.separately=="study"){

      study.names=c("study1", "study2");
      othercovariate.names=NULL;
      event.outcome.names=c("event1", "event2");
      delta.names=c("delta.event1", "delta.event2");
      nonfunctional.covariate.names=c("base_age");
      functional.covariate.names="CAG";
      time.points.for.prediction=  c(seq(40,60,by=1));
      time.points.for.conditional.prediction=c(40, 55);
      time.points.for.conditional.prediction.toadd=c(5,10);
      nonfunctional.covariate.value=c(42);
      functional.covariate.values.of.interest=c(42);

    }


    if(what.analyzed.separately!="none"){
      check.study.equality=FALSE
      estimated.parameters.common.for.all.studies=FALSE
    }else{
      check.study.equality=TRUE
      estimated.parameters.common.for.all.studies=TRUE
    }

    number.of.bootstraps=10;
    use.functional.beta.intercept= TRUE ;
    use.functional.event.coefficients= TRUE;
    use.functional.study.coefficients=TRUE;
    use.bootstrap.variance=TRUE;
    estimation.when.censoring.depends.on.z=FALSE ;
    write.output=FALSE;

#############################################
## a function to estimate JPRAT algorithm  ##
#############################################

  jprat.estimat.results<-jprat.wrapper(study.names=study.names,
                                       input.data.list=input.data.list,
   nonfunctional.covariate.names=nonfunctional.covariate.names,
   functional.covariate.names=functional.covariate.names,
   othercovariate.names=othercovariate.names,
   event.outcome.names=event.outcome.names,
   delta.names=delta.names,
   time.points.for.prediction=time.points.for.prediction,
   time.points.for.conditional.prediction=time.points.for.conditional.prediction,
   time.points.for.conditional.prediction.toadd=time.points.for.conditional.prediction.toadd,
   nonfunctional.covariate.value=nonfunctional.covariate.value,
   functional.covariate.values.of.interest=functional.covariate.values.of.interest,
   number.of.bootstraps=number.of.bootstraps,
   use.functional.beta.intercept=use.functional.beta.intercept,
   use.functional.event.coefficients=use.functional.event.coefficients,
   use.functional.study.coefficients=use.functional.study.coefficients,
   check.study.equality=check.study.equality,
   estimated.parameters.common.for.all.studies=estimated.parameters.common.for.all.studies,
   what.analyzed.separately=what.analyzed.separately,
   estimation.when.censoring.depends.on.z=estimation.when.censoring.depends.on.z,
   use.bootstrap.variance=use.bootstrap.variance, write.output=write.output)

##########################################################
## Get a table for the number of people who are at risk ##
##########################################################
    if(what.analyzed.separately=="event"||what.analyzed.separately=="studyevent"){

      study.names=c("cohort", "predict", "pharos");
      event.outcome.names=c("hdage_nobase", "mcione", "dep2");
      nonfunctional.covariate.names=c("base_age");
      functional.covariate.names="CAG";

    }else if(what.analyzed.separately=="none"||what.analyzed.separately=="study"){

      study.names=c("study1", "study2");
      othercovariate.names=NULL;
      event.outcome.names=c("event1", "event2");
      delta.names=c("delta.event1", "delta.event2");
      nonfunctional.covariate.names=c("base_age");
      functional.covariate.names="CAG";

    }



    if(what.analyzed.separately=="event"||what.analyzed.separately=="studyevent"){

      time.points.for.prediction=seq(46, 66, by=5);
      functional.covariate.values.of.interest=c(46, 48, 50) ;
      nonfunctional.covariate.value=c(40);

    }else if(what.analyzed.separately=="none"||what.analyzed.separately=="study"){

      time.points.for.prediction=  c(seq(40,60,by=1))
      functional.covariate.values.of.interest=c(42);
      nonfunctional.covariate.value=c(42);

    }

    if(what.analyzed.separately!="none"){
      estimated.parameters.common.for.all.studies=FALSE
    }else{
      estimated.parameters.common.for.all.studies=TRUE
    }


## a function to create a table for the number of patients who are at risk
    number.at.risk <- compute.number.at.risk.for.HD(study.names,
     input.data.list,
     event.outcome.names,
     nonfunctional.covariate.names,
     functional.covariate.names,
     nonfunctional.covariate.value,
     functional.covariate.values.of.interest,
     time.points.for.prediction,
     estimated.parameters.common.for.all.studies,
     write.output=FALSE
     )




##############################
## parameters to get results #
##############################

    if(what.analyzed.separately=="event"||what.analyzed.separately=="studyevent"){

      functional.covariate.values.of.interest.ci=c(46, 51);
      time.points.of.interest= c(46, 56);
      time.points.of.interest.ci=seq(46, 66, by=5);
      functional.covariate.comparisons=c(46, 51);
      time.points.for.prediction=seq(46, 66, by=5);
      functional.covariate.values.of.interest=c(46, 48, 50) ;
      nonfunctional.covariate.value=c(40);
      color.names=c("firebrick1", "darkgreen", "black"); ## for color.label.key
      legend.names=c("Motor Diagnosis (DCL=4)") #, "Cognitive Impairment", "Stage II TFC")


    }else{

      functional.covariate.values.of.interest.ci=c(40, 55);
      time.points.of.interest= c(40, 55);
      time.points.of.interest.ci=seq(40, 60, by=5);# c(seq(40,60,by=1))
      functional.covariate.comparisons=c(40, 55);
      time.points.for.prediction=  c(seq(40,60,by=1))
      functional.covariate.values.of.interest=c(42);
      nonfunctional.covariate.value=c(42);
      color.names=c("firebrick1", "darkgreen"); ## for color.label.key
      legend.names=c("event1", "event2"); #c("Motor Diagnosis (DCL=4)", "Cognitive Impairment")

    }


    do.plots=TRUE;
    plot.confidence.intervals=TRUE;
    add.number.at.risk.legend=TRUE;
    ylabel.for.plots.comparing.studies="Probability"
    xlabel.for.plots.comparing.studies="Age (years)"
    file.name.for.analysis="test" ## figure names
    show.results.description=TRUE;

    if(write.output==TRUE){
    is.nrisk.data.frame=TRUE;      ##so is.nrisk.data.frame=TRUE.
                         ## if users choose write.output=FALSE, then nrisk is array form,
                          ## so is.nrisk.data.frame=FALSE.
    write.jprat.output=TRUE; ## users need to choose write.jprat.output=TRUE
                            ## when they choose write.output=TRUE in jprat function;
                            ## FALSE otherwise.
    }else{
      is.nrisk.data.frame=FALSE;
      write.jprat.output=FALSE;
    }

    ####################################
    ## functions to get default values #
    ####################################
    #default.options<-default.options.for.data.setting()

    if(what.analyzed.separately=="studyevent"){


      study.names=c("cohort", "predict", "pharos");
      event.outcome.names=c("hdage_nobase", "mcione", "dep2");
      legend.names=c("Motor Diagnosis (DCL=4)")


    }else if(what.analyzed.separately=="event"){

      study.names=c("cohort", "predict", "pharos");
      event.outcome.names=c("hdage_nobase", "mcione", "dep2");
      legend.names=c("Motor Diagnosis (DCL=4)", "Cognitive Impairment", "Stage II TFC")

    }else if(what.analyzed.separately=="none"||what.analyzed.separately=="study"){

      color.names=c("firebrick1", "darkgreen"); ## for color.label.key
      event.outcome.names=c("event1", "event2")
      legend.names=c("Motor Diagnosis (DCL=4)", "Cognitive Impairment");

    }

    # reformatted.data.for.analysis.results<-data.reformatted.for.analysis.results(
    #   study.names, event.outcome.names,
    #   color.names, legend.names,
    #   which.nonfunctional.covariate.comparisons)

    ## obtain default values
    ## default values
    #use_real_data<-default.options$use_real_data
    #combine.data<-default.options$combine.data

    ## datasets are reformatted, so that results plots and tables can be produced;
    #num_study<-reformatted.data.for.analysis.results$num_study;
    jprat.output<-jprat.estimat.results$jprat.output
    nrisk<-number.at.risk



## a function to display all results including tables and plots
    results.out <- view.all.results(
      ############################
      # arguments for JPRAT     ##
      ############################
      study.names,
      #data.file.names,
      input.data.list,
      nonfunctional.covariate.names,
      functional.covariate.names,
      othercovariate.names,
      event.outcome.names,
      delta.names,
      time.points.for.prediction,
      time.points.for.conditional.prediction,
      time.points.for.conditional.prediction.toadd,
      nonfunctional.covariate.value,
      functional.covariate.values.of.interest,
      #############################
      ## Input: How to analysis  ##
      #############################
      number.of.bootstraps, #=100,
      use.functional.beta.intercept,
      use.functional.event.coefficients,
      use.functional.study.coefficients,
      check.study.equality,
      estimated.parameters.common.for.all.studies,
      what.analyzed.separately,
      estimation.when.censoring.depends.on.z,
      use.bootstrap.variance,
      ##########################
      ## For plotting results ##
      ##########################
      functional.covariate.values.of.interest.ci,
      time.points.of.interest,
      time.points.of.interest.ci,
      label.for.alpha.values.over.time=NULL,
      #which.nonfunctional.covariate.comparisons,
      color.names=color.names, ## returns color.labels
      legend.names=legend.names, ## returns legend.labels
      functional.covariate.comparisons=functional.covariate.comparisons,
      functional.covariate.comparisons.for.sample.size=NULL,
      #############################
      # plot options             ##
      #############################
      do.plots=do.plots,
      plot.confidence.intervals=plot.confidence.intervals,
      ###########################
      # Label figures or files ##
      ###########################
      add.number.at.risk.legend=add.number.at.risk.legend,
      ylabel.for.plots.comparing.studies=ylabel.for.plots.comparing.studies,
      xlabel.for.plots.comparing.studies=xlabel.for.plots.comparing.studies,
      file.name.for.analysis=file.name.for.analysis, ## figure names
      #######################
      ## Output from JPRAT ##
      #######################
      #data.truth=data.truth,
      #data.theta=data.theta,
      #data.combi=data.combi,
      jprat.output=jprat.output,
      #data.count=data.count,
      #data.count.outside=data.count.outside,
      nrisk=nrisk,
      #####################################
      ## Do we show results description? ##
      #####################################
      show.results.description=show.results.description,
      is.nrisk.data.frame=is.nrisk.data.frame,
      write.jprat.output=write.jprat.output
    )


}
\references{
To see the detailed description for the JPRAT estimation procedure, we refer to
         \itemize{
         \item Garcia, T. P., Marder, K., Wang, Y. (2017). Time-varying proportional odds model for mega-analysis of clustered event
         times. Biostatistics, 20(1), 129-146.
         }
         To see the detailed real data analysis, we refer to
         \itemize{
         \item Garcia, T.P., Wang, Y., Shoulson, I., Paulsen, J.S. and Marder, K. (2018). Disease progression in Huntington disease:
         An analysis of multiple longitudinal outcomes. Journal of Huntington's disease, 7, 337-344}
}
