% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{view.all.results}
\alias{view.all.results}
\title{This function returns analysis results including plots and tables.}
\usage{
view.all.results(
  study.names,
  input.data.list,
  nonfunctional.covariate.names,
  functional.covariate.names,
  othercovariate.names,
  event.outcome.names,
  delta.names,
  time.points.for.prediction,
  time.points.for.conditional.prediction,
  time.points.for.conditional.prediction.toadd,
  nonfunctional.covariate.value,
  functional.covariate.values.of.interest,
  number.of.bootstraps,
  use.functional.beta.intercept,
  use.functional.event.coefficients,
  use.functional.study.coefficients,
  check.study.equality,
  estimated.parameters.common.for.all.studies,
  what.analyzed.separately,
  estimation.when.censoring.depends.on.z,
  use.bootstrap.variance,
  functional.covariate.values.of.interest.ci,
  time.points.of.interest,
  time.points.of.interest.ci,
  label.for.alpha.values.over.time = NULL,
  which.nonfunctional.covariate.comparisons,
  color.names,
  legend.names,
  functional.covariate.comparisons,
  functional.covariate.comparisons.for.sample.size = NULL,
  do.plots,
  plot.confidence.intervals,
  add.number.at.risk.legend,
  ylabel.for.plots.comparing.studies,
  xlabel.for.plots.comparing.studies,
  file.name.for.analysis,
  data.theta = data.theta,
  data.combi = data.combi,
  data.count = data.count,
  data.count.outside = data.count.outside,
  nrisk = nrisk,
  data.truth.other = NULL,
  data.theta.other = NULL,
  nrisk.other = NULL,
  data.combi.other = NULL,
  s.names.use = NULL,
  s.names.other.use = NULL,
  type1.error = NULL,
  type2.error = NULL,
  treatment.effect = NULL,
  dropout.rate = NULL,
  show.results.description
)
}
\arguments{
\item{study.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{input.data.list}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{othercovariate.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{event.outcome.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{delta.names}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.prediction}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.conditional.prediction}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{time.points.for.conditional.prediction.toadd}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{nonfunctional.covariate.value}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{number.of.bootstraps}{ADD DETAILS HERE!}

\item{use.functional.beta.intercept}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{use.functional.event.coefficients}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{use.functional.study.coefficients}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{check.study.equality}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{estimated.parameters.common.for.all.studies}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{what.analyzed.separately}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{estimation.when.censoring.depends.on.z}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{use.bootstrap.variance}{See this argument in the \code{\link{jprat.wrapper}} function.}

\item{functional.covariate.values.of.interest.ci}{a vector of specific functional covariate values of X,
 where the confidence interval of smooth functional parameter \eqn{\alpha(X=x,t)} will be estimated.
 For example, we set the values \eqn{x=46, 51} in the analysis.
 We recommend users to choose the same vector of functional covariate values X as in the interested functional covariate values (\code{functional.covariate.values.of.interest})
to ease estimation procedure in the analysis.}

\item{time.points.of.interest}{a vector of specific time points at which the smooth functional parameters \eqn{\alpha(X,t)}
will be predicted. These time points are used to be labeled for the plot.
For example, we set the values \eqn{x=46, 56} in the analysis.
We recommend users to choose these time points vector as the subset of time points for prediction (\code{time.points.for.prediction}).}

\item{time.points.of.interest.ci}{a vector of time points at which confidence intervals of \eqn{\alpha(X,t)} are predicted.
These time points are used to be labeled for the plot. For example, we set the values \eqn{x=46, 51, 56, 61, 66} in the analysis.
We recommend that these time points vector should be chosen as the subset of time points for prediction (\code{time.points.for.prediction}) or as the same time points for estimating \eqn{\alpha(X,t)} (\code{time.points.of.interest}).}

\item{label.for.alpha.values.over.time}{a vector of specific time points at which the smooth functional parameters \eqn{\alpha(X,t)} will be predicted.
These time points are used to be labeled for the plot. Currently, we do not make a plot for the smooth functional parameters \eqn{\alpha(X,t)}.
Default is NULL.}

\item{which.nonfunctional.covariate.comparisons}{a numeric vector for which nonfunctional covariates will be compared. In our analysis, we only consider one nonfunctional covariate named "base_age", default is c(1, 1).
Currently, we only consider one nonfunctional covariate and there is nothing to compare between covariates.}

\item{color.names}{a character vector to denote which colors will be used for the labels of events of interest in plots. e.g., c("firebrick1", "darkgreen", "black").}

\item{legend.names}{a character vector to label event of interests in plots. e.g., c("Motor Diagnosis (DCL=4)",  "Cognitive Impairment" , "Stage II TFC").}

\item{functional.covariate.comparisons}{a vector of functional covariate values of X at which the smooth functional parameters \eqn{\alpha(X=x,\cdot)} were estimated and compared.
For example, the functional covariate values are \eqn{x=c(46, 51)} and the parameters of \eqn{\alpha(x=46,\cdot)} and \eqn{\alpha(x=51,\cdot)} will be compared.
We recommend that these covariate values X should be the same as the vector of functional covariate values of X for confidence intervals (\code{functional.covariate.values.of.interest.ci}).}

\item{functional.covariate.comparisons.for.sample.size}{a specific functional covariate values of X to use the sample size calculation,
which will be compared across studies. Currently, we are not interested in sample size calculation. Default is NULL.}

\item{do.plots}{a logical value whether the results plot will be generated. Default is TRUE.}

\item{plot.confidence.intervals}{a logical value whether the confidence intervals of the predicted marginal distributions will be shown in the plot. Default is TRUE.}

\item{add.number.at.risk.legend}{a logical value whether number of subjects who is at risks at the interested time points will be added with legends in plots.}

\item{ylabel.for.plots.comparing.studies}{a character value for y-label in plots when studies are comparing or event of interests were compared in each study. Default is "Probability".}

\item{xlabel.for.plots.comparing.studies}{a character value for x-label in plots when studies are comparing for each event of interest or event of interests were compared in each study. Default is "Age(years)".}

\item{file.name.for.analysis}{a character value, which is used as the part of plot names.}

\item{data.theta}{data file for the estimated values of all components as a result of analysis. The result file is automatically created if users choose \code{write.output=TRUE}.
See return value \code{theta.out} in the \code{\link{jprat.wrapper}} function.}

\item{data.combi}{data file for the difference between the estimated values of all components among studies as a result of analysis.
Default is data.combi and this file is automatically created if users choose \code{write.output=TRUE}.
See return value \code{combi.out} in the \code{\link{jprat.wrapper}} function.}

\item{data.count}{data file for counts for binary status of censored, uncensored and missing subjects as a result of analysis.
Default is data.count and this file is automatically created if users choose \code{write.output=TRUE}. See return value \code{count.store} in the \code{\link{jprat.wrapper}} function.}

\item{data.count.outside}{data file for counts for binary status of censored, uncensored outside [0,1] as a result of analysis. Default is \code{data.count.outside} and this file is automatically created if users choose \code{write.output=TRUE}.
See return value \code{count.store.outside} in the \code{\link{jprat.wrapper}} function.}

\item{nrisk}{data file for the number of subjects who is at risk at the interested the time points defined by users. This is automatically calculated.
Default is nrisk and this file is automatically created.}

\item{data.truth.other}{This is not a users' interest argument. Analysis results from COX model. Use this if the proportional odds model is compared to COX model, which is not considered. Default is NULL.}

\item{data.theta.other}{This is not a users' interest argument. Analysis results from COX model. Use this if the proportional odds model is compared to COX model, which is not considered. Default is NULL.}

\item{nrisk.other}{This is not a users' interest argument. Analysis results from COX model. Use this if the proportional odds model is compared to COX model, which is not considered. Default is NULL.}

\item{data.combi.other}{This is not a users' interest argument. Analysis results from COX model. Use this if the proportional odds model is compared to COX model, which is not considered. Default is NULL.}

\item{s.names.use}{This is not a users' interest argument. Default is NULL.}

\item{s.names.other.use}{This is not a users' interest argument. Default is NULL.}

\item{type1.error}{type 1 error rate, which is for sample size calculations. We do not consider this arguments in the analysis. Default is NULL.}

\item{type2.error}{type 2 error rate, which is for sample size calculations. We do not consider this arguments in the analysis. Default is NULL.}

\item{treatment.effect}{a numeric vector of the treatment effects for sample size calculations. We do not consider this arguments in the analysis. We do not consider this arguments in the analysis. Default is NULL.}

\item{dropout.rate}{a numeric vector of drop-out rates, which is used for sample size calculations. We do not consider this arguments in the analysis. Default is NULL.}

\item{show.results.description}{a logical value whether the description of plots and tables will be produced. Default is FALSE.}
}
\value{
This function returns ggplots.
}
\description{
This function returns analysis results including plots and tables.
}
\details{
The number of risks at each time point will be printed in the plot.
}
\examples{

## load your data in the list format: should match order of study.names
input.data.list=list(cohort=data_cohort, predict=data_predict, pharos=data_pharos);

#Input data for JPRAT estimation procedure
study.names=c("cohort", "predict", "pharos");
nonfunctional.covariate.names=c("base_age");
functional.covariate.names="CAG";
othercovariate.names=c("firstyear", "lastyear")
event.outcome.names=c("hdage_nobase", "mcione", "tfctwo");
delta.names=c("delta.hdage_nobase", "delta.mcione", "delta.tfctwo");
time.points.for.prediction=seq(46, 66, by=5);
time.points.for.conditional.prediction=c(46, 51, 56);
time.points.for.conditional.prediction.toadd=c(5);
nonfunctional.covariate.value=c(40);
functional.covariate.values.of.interest=c(46, 48, 50) ;

# parameter setup to analyze data using JPRAT estimation procedure
number.of.bootstraps=100;
use.functional.beta.intercept= TRUE ;
use.functional.event.coefficients= TRUE;
use.functional.study.coefficients=TRUE;
check.study.equality=FALSE;
what.analyzed.separately="studyevent";
estimated.parameters.common.for.all.studies=FALSE;
use.bootstrap.variance=TRUE ;
estimation.when.censoring.depends.on.z=FALSE ;
write.output=TRUE;

# JPRAT estimation
jprat.estimat.results<-jprat.wrapper(study.names=study.names,
                                     input.data.list=input.data.list,
 nonfunctional.covariate.names=nonfunctional.covariate.names,
 functional.covariate.names=functional.covariate.names,
 othercovariate.names=othercovariate.names,
 event.outcome.names=event.outcome.names,
 delta.names=delta.names,
 time.points.for.prediction=time.points.for.prediction,
 time.points.for.conditional.prediction=time.points.for.conditional.prediction,
 time.points.for.conditional.prediction.toadd=time.points.for.conditional.prediction.toadd,
 nonfunctional.covariate.value=nonfunctional.covariate.value,
 functional.covariate.values.of.interest=functional.covariate.values.of.interest,
 number.of.bootstraps=number.of.bootstraps,
 use.functional.beta.intercept=use.functional.beta.intercept,
 use.functional.event.coefficients=use.functional.event.coefficients,
 use.functional.study.coefficients=use.functional.study.coefficients,
 check.study.equality=check.study.equality,
 estimated.parameters.common.for.all.studies=estimated.parameters.common.for.all.studies,
 what.analyzed.separately=what.analyzed.separately,
 estimation.when.censoring.depends.on.z=estimation.when.censoring.depends.on.z,
 use.bootstrap.variance=use.bootstrap.variance, write.output=write.output)


##get number at risk table
study.names=c("cohort", "predict", "pharos");
input.data.list=list(cohort=data_cohort, predict=data_predict, pharos=data_pharos);
event.outcome.names=c("hdage_nobase", "mcione", "tfctwo");
nonfunctional.covariate.names=c("base_age");
functional.covariate.names="CAG";
nonfunctional.covariate.value=c(40);
functional.covariate.values.of.interest=c(46, 48, 50) ;
time.points.for.prediction=seq(46, 66, by=5);
estimated.parameters.common.for.all.studies=FALSE;

##to creat out_nrisk.dat file
number.at.risk <- compute.number.at.risk.for.HD(study.names,
 input.data.list, #data.file.names,
 event.outcome.names,
 nonfunctional.covariate.names,
 functional.covariate.names,
 nonfunctional.covariate.value,
 functional.covariate.values.of.interest,
 time.points.for.prediction,
 estimated.parameters.common.for.all.studies
 )



## To get results
#############################
# parameters to get results #
#############################
functional.covariate.values.of.interest.ci=c(46, 51);
time.points.of.interest= c(46, 56);
time.points.of.interest.ci=seq(46, 66, by=5);
which.nonfunctional.covariate.comparisons=c(1,1);
color.names=c("firebrick1", "darkgreen", "black"); ## for color.label.key
legend.names=c("Motor Diagnosis (DCL=4)", "Cognitive Impairment", "Stage II TFC")
functional.covariate.comparisons=c(46, 51);
do.plots=TRUE;
plot.confidence.intervals=TRUE;
add.number.at.risk.legend=TRUE;
ylabel.for.plots.comparing.studies="Probability"
xlabel.for.plots.comparing.studies="Age (years)"
file.name.for.analysis="test" ## figure names
show.results.description=FALSE;

####################################
## functions to get default values #
####################################
default.options<-default.options.for.data.setting()

## data is reformatted to get analysis results
reformatted.data.for.analysis.results<-data.reformatted.for.analysis.results(
  study.names, event.outcome.names,
  color.names, legend.names,
  which.nonfunctional.covariate.comparisons)

## obtain default values
## default values
use_real_data<-default.options$use_real_data
combine.data<-default.options$combine.data

## datasets are reformatted, so that results plots and tables can be produced;
num_study<-reformatted.data.for.analysis.results$num_study;

## load results file from JPRAT algorithm
#data.truth <- read.table("out_truth_real_output_iseed_1.dat",header=FALSE)
data.theta <- read.table("out_thetaest_real_output_iseed_1.dat",header=FALSE)

if(num_study > 1){
  data.combi <- read.table("out_combi_real_output_iseed_1.dat",header=FALSE)
} else {
  data.combi <- NULL
}

data.count <- read.table("out_count_real_output_iseed_1.dat",header=FALSE)
data.count.outside <- read.table("out_outside_count_real_output_iseed_1.dat",header=FALSE)
nrisk <- read.table("out_nrisk.dat",header=TRUE)


## display all results including tables and plots
results.out <- view.all.results(
  ############################
  # arguments for JPRAT     ##
  ############################
  study.names,
  #data.file.names,
  input.data.list,
  nonfunctional.covariate.names,
  functional.covariate.names,
  othercovariate.names,
  event.outcome.names,
  delta.names,
  time.points.for.prediction,
  time.points.for.conditional.prediction,
  time.points.for.conditional.prediction.toadd,
  nonfunctional.covariate.value,
  functional.covariate.values.of.interest,
  #############################
  ## Input: How to analysis  ##
  #############################
  number.of.bootstraps, #=100,
  use.functional.beta.intercept,
  use.functional.event.coefficients,
  use.functional.study.coefficients,
  check.study.equality,
  estimated.parameters.common.for.all.studies,
  what.analyzed.separately,
  estimation.when.censoring.depends.on.z,
  use.bootstrap.variance,
  ##########################
  ## For plotting results ##
  ##########################
  functional.covariate.values.of.interest.ci,
  time.points.of.interest,
  time.points.of.interest.ci,
  label.for.alpha.values.over.time=NULL,
  which.nonfunctional.covariate.comparisons,
  color.names, ## returns color.labels
  legend.names, ## returns legend.labels
  functional.covariate.comparisons,
  functional.covariate.comparisons.for.sample.size=NULL,
  #############################
  # plot options             ##
  #############################
  do.plots,
  plot.confidence.intervals,
  ###########################
  # Label figures or files ##
  ###########################
  add.number.at.risk.legend,
  ylabel.for.plots.comparing.studies,
  xlabel.for.plots.comparing.studies,
  file.name.for.analysis, ## figure names
  #######################
  ## Output from JPRAT ##
  #######################
  #data.truth=data.truth,
  data.theta=data.theta,
  data.combi=data.combi,
  data.count=data.count,
  data.count.outside=data.count.outside,
  nrisk=nrisk,
  data.truth.other=NULL,
  data.theta.other=NULL,
  nrisk.other=NULL,
  data.combi.other=NULL,
  s.names.use=NULL,
  s.names.other.use=NULL,
  ########################################################
  ## for computing sample sizes: users do not need them ##
  ########################################################
  type1.error=NULL,
  type2.error=NULL,
  treatment.effect=NULL,
  dropout.rate=NULL,
  #####################################
  ## Do we show results description? ##
  #####################################
  show.results.description
)


}
